refactor(mcp): extract McpToolDetailDialog to separate file #371
Move dialog implementation from McpToolDetailPanel to its own file for better code organization and maintainability. Also improves description display with JTextPane.
